Hair - creating parent class for raster pass as prep for ShortCut technique
